Changes to support 64 bit
git-svn-id: https://svn.code.sf.net/p/directshownet/code/trunk@1033 b2a09c84-0b36-484d-b36e-ffab392acda3

+++ directshowlib/Test/v1.0/IMediaEventTest.cs
@@ -28,7 +28,7 @@ namespace DirectShowLib.Test
         {
             int hr;
             IntPtr hEvent;
-            int p1, p2;
+            IntPtr p1, p2;
             EventCode ec;
 
             BuildGraph();
