+++ directshowlib/src/Control.cs
@@ -425,8 +425,8 @@ namespace DirectShowLib
         int NotifyOwnerMessage(
             [In] IntPtr hwnd, // HWND *
             [In] int msg,
-            [In] int wParam, // WPARAM
-            [In] int lParam // LPARAM
+            [In] IntPtr wParam, // WPARAM
+            [In] IntPtr lParam // LPARAM
             );
 
         [PreserveSig]
@@ -777,8 +777,8 @@ namespace DirectShowLib
         [PreserveSig]
         int GetEvent(
             [Out] out EventCode lEventCode,
-            [Out] out int lParam1,
-            [Out] out int lParam2,
+            [Out] out IntPtr lParam1,
+            [Out] out IntPtr lParam2,
             [In] int msTimeout
             );
 
@@ -797,8 +797,8 @@ namespace DirectShowLib
         [PreserveSig]
         int FreeEventParams(
             [In] EventCode lEvCode,
-            [In] int lParam1,
-            [In] int lParam2
+            [In] IntPtr lParam1,
+            [In] IntPtr lParam2
             );
     }
 
@@ -815,8 +815,8 @@ namespace DirectShowLib
         [PreserveSig]
         new int GetEvent(
             [Out] out EventCode lEventCode,
-            [Out] out int lParam1,
-            [Out] out int lParam2,
+            [Out] out IntPtr lParam1,
+            [Out] out IntPtr lParam2,
             [In] int msTimeout
             );
 
@@ -835,8 +835,8 @@ namespace DirectShowLib
         [PreserveSig]
         new int FreeEventParams(
             [In] EventCode lEvCode,
-            [In] int lParam1,
-            [In] int lParam2
+            [In] IntPtr lParam1,
+            [In] IntPtr lParam2
             );
 
         #endregion
